REFACTOR: change schedules route structure
