--- src/app/(home)/@detail/(.)schedules/[sid]/page.tsx
@@ -1,6 +1,0 @@
-import ScheduleDetail from "@/app/components/schedule/detail";
-
-export default async function Page({ params }: { params: Promise<{ sid: string }> }) {
-  const { sid } = await params;
-  return <ScheduleDetail scheduleId={Number(sid)} />;
-}
